fix: move to plot mixin
